The goal of this work was to propose a methodology to determine the smallest sample size that captures degradation in a cell accounting the cell-to-cell variation in a larger population  Study reproduced: Strange, Calum, et al. “Automatic Method for the Estimation of Li-Ion Degradation Test Sample Sizes Required to Understand Cell-to-Cell Variability.” Energy and AI, vol. 9, 2022, p. 100174., https://doi.org/10.1016/j.egyai.2022.100174.
The goal of this work was to propose a methodology to determine the smallest sample size that captures degradation in a cell accounting the cell-to-cell variation in a larger population

Study reproduced: Strange, Calum, et al. “Automatic Method for the Estimation of Li-Ion Degradation Test Sample Sizes Required to Understand Cell-to-Cell Variability.” Energy and AI, vol. 9, 2022, p. 100174., https://doi.org/10.1016/j.egyai.2022.100174.